+++ src/sort/merge_sort.rs
@@ -9,8 +9,8 @@ fn merge<T: PartialOrd + Copy>(
     index_left: usize,
     index_right: usize,
     index_middle: usize,
-    comparisons: &mut usize,
-    moves: &mut usize,
+    comparisons: &mut u64,
+    moves: &mut u64,
     pb: &ProgressBar,
 ) {
     let left_array_len: usize = index_middle - index_left + 1;
@@ -73,8 +73,8 @@ fn merge_sort_recursive<T: PartialOrd + Copy>(
     array: &mut Vec<T>,
     index_left: usize,
     index_right: usize,
-    comparisons: &mut usize,
-    moves: &mut usize,
+    comparisons: &mut u64,
+    moves: &mut u64,
     pb: &ProgressBar,
 ) {
     if index_left < index_right {
@@ -104,12 +104,12 @@ pub fn sort<T: PartialOrd + Copy>(mut array: Vec<T>) -> SortResult<T> {
         .progress_chars("#>-"),
     );
 
-    let mut comparisons: usize = 0;
-    let swaps: usize = 0;
-    let shifts: usize = 0;
-    let insertions: usize = 0;
-    let mut moves: usize = 0;
-    let attempts: usize = 0;
+    let mut comparisons: u64 = 0;
+    let swaps: u64 = 0;
+    let shifts: u64 = 0;
+    let insertions: u64 = 0;
+    let mut moves: u64 = 0;
+    let attempts: u64 = 0;
 
     let start = Instant::now();
 
@@ -124,7 +124,7 @@ pub fn sort<T: PartialOrd + Copy>(mut array: Vec<T>) -> SortResult<T> {
         &pb,
     );
 
-    let duration: usize = start.elapsed().as_nanos() as usize;
+    let duration: u128 = start.elapsed().as_nanos();
     pb.finish_with_message("Sorting completed!");
 
     SortResult {
@@ -146,10 +146,10 @@ mod tests {
 
     #[test]
     fn sort_array() {
-        let array: Vec<i32> = (1..=10).rev().collect();
+        let array: Vec<u64> = (1..=10).rev().collect();
 
         let result = sort(array);
-        let expected: Vec<i32> = (1..=10).collect();
+        let expected: Vec<u64> = (1..=10).collect();
 
         assert_eq!(
             result.array, expected,
